+++ src/config.lua
@@ -1,6 +1,8 @@
 ---@meta KeepsakeChambers-zannc-config
 return {
+	enabled = true,
 	Increment = 5,
 }, {
+	enabled = "Set to true to enable the mod, set to false to disable it.",
 	Increment = "Amount of 'chamber clears' to increment by on a chamber clear, eg increment of 5 will get to 25 chambers in 5 rooms.\nIncrement of 5 allows you to max a keepsake in 15 rooms",
 }