+++ src/ready.lua
@@ -1,6 +1,0 @@
----@meta _
----@diagnostic disable: lowercase-global
-
-modutil.mod.Path.Context.Wrap("AdvanceKeepsake", function()
-	return AdvanceKeepsake_wrap()
-end)